Try harder to make file locking succeed in case of network errors
We noticed that during nightly builds, the LockFileEx() call would
occasionally(!) fail with ERROR_NETNAME_DELETED. The lockfile was
specified via an UNC path to a network drive, but the host was
available all the time. We couldn't find a reason why this would
happen and there was no obvious pattern to this either.

Let's just try to call LockFileEx() again in case this error occurs
(at least a couple of times), it might be that it's just a temporary
failure.